backend/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
psycopg2-binary==2.9.9
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
supabase==2.0.3
minio==7.1.15
python-keycloak==3.3.0
PyJWT==2.8.0
requests==2.31.0
